+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### vNEXT
+* `ExpressApollo` exposes the response object to the server's options callback.
+
 ### v0.2.0
 * Complete refactor of Apollo Server using TypeScript. PR [#41](https://github.com/apollostack/apollo-server/pull/41) including the following changes:
 * Dropped express-graphql dependency